requirements.txt
python-dotenv
requests
groq
PyPDF2